refactor: move App in a folder
